Add pubsys-setup to ease creation of root role and signing key
If the user doesn't define a root role and key themselves in Infra.toml, we'll
generate them so it's easier to get started building images and repos.  The
user should update Infra.toml with more permanent resources and locations, but
this makes it much easier to get started handling your own builds.

Note: this changes the format of Infra.toml to make root roles and keys
specific to a repo.  Roles and keys are necessarily tied to a repo, so this
is simpler than having to specify PUBLISH_REPO *and* PUBLISH_KEY (and manually
changing the root.json file, which wasn't configurable before).

Co-authored-by: Zac Mrowicki <mrowicki@amazon.com>
Co-authored-by: Tom Kirchner <tjk@amazon.com>