model: fix validity issues
Constraints on the way that OpenAPI is used with Kubernetes CRD objects
prohibits algebraic enum variants. To fix this and make the CRD valid
we needed to move the error message to its own field.

Additionally, when a CRD object has no status, and we attempt to patch
just one field in the status, then it is necessary for the other fields
in the status to be optional.