controller: remove duplicate trace
The controller was tracing at the start of reconciliation loop and at
the end of the loop. Remove the one at the end of the loop to reduce
useless log lines.

Also set the default log level to Trace in the controller because we are
almost always going to want this while we are working on development.
When we are getting closer to being done we can set it to Info or Debug.