test-agent: simplify runner trait
Simplify the Runner trait so that it returns from a run function once
tests are complete instead of running them in a background thread (or
process). Also, eliminate timeouts for now. The controller can probably
handle this.

Both tests are also simplified to use println! instead of shelling out
to echo.