build: re-enable ssl
In 5e8279e we accidentally eliminated all SSL implementations from the
test-agent and resource-agents builds which lead to an obscure failure
mode (Connection reset by peer os error 104). Here we enable the
native-tls feature to fix this problem.

Doing so also revealed that the use of Cargo Resolver = 2 as a
workaround for incorrect flags during build.rs compilation was not fully
correct. To fix it we make sure that yamlgen only ever shows up as a
[build-dependency] (and never a [dependency] or [dev-dependency]).