build: rename arch variables
I spent a bunch of time figuring out why the Sonobuoy build was not
working and it turned out to be because some other script of mine had
put the variable ARCH into my environment. Here I rename the Makefile
variables UNAME_ARCH and ARCH to TESTSYS_BUILD_HOST_UNAME_ARCH and
TESTSYS_BUILD_HOST_GOARCH respectively.

GOARCH is used to reflect that the less-common amd64 and arm64
designations exist in our build because aws-iam-authenticator is a Go
program (and presumably this is why they use the GOARCH values when
publishing).

I also added a printout of these variables to image builds to hopefully
help future me when our build is broken again for some future reason.