{eks,ec2}-resource-agent, testsys: generate cluster-dns-ip setting for BR userdata
The eks-resource-agent now determines the cluster-dns-ip from cluster
info and sets it in the cluster output info.
The ec2-resource-agent takes the cluster-dns-ip from the K8s cluster
info and specifies the cluster-dns-ip setting in Bottlerocket's userdata
when launching the nodes.

All of this allows Bottlerocket to set its K8s node IP to the
appropriate internal IP (either IPv4 address or IPv6 address)