Add help target to Makefile
This adds a `help` target and sets it as the default. This helps newer
contributors discover the make targets of relevance by providing a
nicely formatted output of the targets they might be interested in.

This can be expanded as we go to add additional targets or to categorize
targets into different groups based on their expected usage.

Also some minor cleanup to the Makefile itself.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>