Add Eq trait to model structs
With the update to the latest ructc release (1.63.0) there was a new
error introduced when running clippy checks. We had a couple structs in
our model that had the `PartialEq` trait, but could also have `Eq` since
all of their members supported `Eq`.

This updates the two structs with the `Eq` trait. Once past that error
there was also a new error about an unnecessary use of `return`. So this
removes that statement and reformats that section of code to be
consistent with the similar code block above it.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>