CI: Skip some tasks on push
Several of our CI tests run on both pull_request (when a PR is
submitted) and push (when the PR is actually merged into the develop
branch). Some things make sense to run on both as a safeguard to make
sure multiple changes in flight did not cause an issue. But some things
will not have changed.

This makes the "license check" task only run on the PR as nothing in
licensing should change by the time a PR is committed into the branch.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>