Actions: update docs directory path
We originally had a `design` directory that contained the design doc for
testsys. We added more docs and restructured things, so `design` was
renamed to `docs`. There is a GitHub Action configuration that tries to
ignore this path to skip running tests if only docs have changed. This
updates the config to use the correct current name.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>