cleanup: Correct sonobuoy test yaml field names
Since we use `#[serde(rename_all = "camelCase")]` in the configuration
sruct, the Rust native names are changed from snake case to camel case
in the result serialized YAML.

The example YAML shown in the comments of the sonobuoy-test-agent code
did not make this translation, so if the example is copy/pasted and
attempted to be used as-is, the deserialization to a CRD fails because
of missing and mismatched fields.

This updates the example YAML to be the correct modified names.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>