sonobuoy: Fix regression while retrieving progress
Sometimes (always?) the JSON results of a sonobuoy plugin run will have
a "progress" field with additional information that an object rather
than the string. We were only handling the string case, causing an error
when we tried to parse the object results.

This updates our parsing to treat whatever is present in the `progress`
field as its string representation. Added a new unit test to  cover this
scenario and updates existing tests to ensure we handle either input
correctly.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>
Co-authored-by: Ethan Pullen <pullenep@amazon.com>