Add NVIDIA workload test definition
This adds a new `tests` directory tree to contain our workload tests and
any other type of add-on test type we decide to add in the future. It
adds our first workload test for running NVIDIA smoketests. Basic
instructions are included to help guide future test authors of the
requirements for tests and where they should be placed.

Some of this content will need to be updated once we fully finish
publishing the container images and flesh out things like ECS testing
and how the full execution of workload tests will be performed. This
gets the majority of the pieces in place so those updates will be
smaller and it will be easier to complete that work using published
content.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>