controller: Resource creation waits for test
Resources now wait for a test that requires them before they are
created. This along with an addition to `is_delete_required()` that
checks to make sure the resource exists before deleting it solve a race
condition where a resourcee that was applied to a cluster slightly
before its corresponding test would automatically be cleaned up before
it was created.