Remove workaround for eksctl build
There was a release of eksctl that would fail due to a retagged
dependency causing a hash mismatch. To get past this we added some
handling to update the go.mod/go.sum files when building eksctl as part
of our tools build.

eksctl has now moved past that broken version of the dependency, and the
current version we use no longer needs this workaround, so this cleans
up that extra handling.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>