Remove extraneous `default()` calls
Starting with Rust 1.71.0, clippy had added another check that looks for
`default` being called on unit structs. There are a few instances in the
testsys codebase where this happens, so attempting to build with the
latest rust toolchain will fail.

This removes the calls to `default()` since they are not necessary.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>