agents: Build eksctl-anywhere from source
Existing method gets eksctl-anywhere by curling the binary of a specific
release version. Base functionality is fairly stable, and we often need
to pick up the latest when adding a new variant to get support for a
newer k8s version. That is the case now with needing the k8s 1.28 support
that is not in a release yet.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>