metal,vsphere-cluster-resource: retrieve EKS-A binary at runtime
Adds a new field for specifying the EKS-A release manifest URL to both
the metal and vSphere cluster resource agent configurations. At runtime,
the agents will download the EKS-A binary archive tagged as the latest
release in the EKS-A release manifest. If a manifest is not specified,
then the agents defaults to using the upstream official EKS-A release
manifest.