fix: Add new way to create cluster, create cluster from yaml file, to fix previously wrong ipFamily issue
This is the body of the commit: this fix addressed the issue of when
create clusters, the ipv6 cluster always has a ipFamily of ipv4 which is
wrong. In this commit, a new way of creating cluster is added which is
using created yaml file in a certain path. That way we can definite the
ipFamily for whether it is ipv6 or ipv4.

Fixed feedbacks

fixed build